refactor: Simplify tools folder structure
