--- tools/11ty/root/sitemap.xml
@@ -1,7 +1,0 @@
-<?xml version='1.0' encoding='UTF-8'?>
-<!--  generator="Inspyder Sitemap Creator 4.1.3"  -->
-<urlset xmlns="https://www.sitemaps.org/schemas/sitemap/0.9" xmlns:xsi="https://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="https://www.sitemaps.org/schemas/sitemap/0.9 https://www.sitemaps.org/schemas/sitemap/0.9/sitemap.xsd">
-<!--  your website  pages URLs go here"  -->
-<url><loc>https://yoursite.url</loc></url>
-
-</urlset>
