--- tools/11ty/root/web.config
@@ -1,29 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<configuration>
-  <system.webServer>
-    <staticContent>
-      <remove fileExtension=".woff" />
-      <!-- in case the mime type already exist in IIS -->
-      <clientCache cacheControlCustom="public" cacheControlMode="UseMaxAge" cacheControlMaxAge="97.00:00:00" />
-      <mimeMap fileExtension=".woff" mimeType="application/x-font-woff" />
-      <mimeMap fileExtension=".woff2" mimeType="application/font-woff2" />
-    </staticContent>
-    <!-- Below block is for custom redirect for 404 error to 404.html -->
-    <httpErrors>
-      <remove statusCode="404" />
-      <error statusCode="404" path="404.html" />
-    </httpErrors>
-    <rewrite>
-      <rules>
-        <rule name="Production Robots">
-          <match url="robots.txt" />
-          <conditions logicalGrouping="MatchAny">
-            <add input="{HTTP_X_FORWARDED_HOST}" pattern="^www\.mydomain\.ca\.gov$" />
-            <add input="{HTTP_HOST}" pattern="^www\.mydomain\.ca\.gov$" />
-          </conditions>
-          <action type="Rewrite" url="robots-prod.txt" />
-        </rule>
-      </rules>
-    </rewrite>
-  </system.webServer>
-</configuration>
