+++ tools/bundlers/run-bundlers.js
@@ -1,5 +1,5 @@
-import cssBuilder from "../bundlers/cssBuilder.js";
-import jsBuilder from "../bundlers/jsBuilder.js";
+import cssBuilder from "./css-builder.js";
+import jsBuilder from "./js-builder.js";
 
 await cssBuilder("dist/bundle.css");
 await cssBuilder("dist/bundle.min.css", { minify: true });
